Compile with type checking
