--- .babelrc
@@ -1,7 +1,0 @@
-{
-    "presets": [["es2015", {"modules": false}], "react"],
-    "plugins": [
-        ["transform-class-properties", { "spec": true }],
-        ["transform-object-rest-spread"]
-    ]
-}
