+++ package.json
@@ -32,7 +32,8 @@
     "author": "Maxime De Waegeneer <mdewaegeneer@gmail.com>",
     "license": "GPL-3.0",
     "dependencies": {
-        "@types/google-protobuf": "^3.2.7",
+        "@improbable-eng/grpc-web": "^0.12.0",
+        "@types/google-protobuf": "^3.7.2",
         "appdirs": "^1.1.0",
         "bitlyapi": "^1.0.9",
         "command-exists": "^1.2.6",
@@ -47,7 +48,6 @@
         "grpc-bus-websocket-server": "file:opt/scopeserver/bindserver",
         "grpc-caller": "^0.3.1",
         "grpc-tools": "^1.6.6",
-        "grpc-web-client": "^0.3.1",
         "json-schema": "^0.2.3",
         "json2csv": "^4.0.1",
         "launch-browser": "^1.0.0",
@@ -88,10 +88,14 @@
         "@babel/cli": "^7.8.4",
         "@babel/core": "^7.8.7",
         "@babel/plugin-proposal-class-properties": "^7.8.3",
+        "@babel/plugin-proposal-object-rest-spread": "^7.8.3",
+        "@babel/preset-env": "^7.8.7",
         "@babel/preset-react": "^7.8.3",
         "@babel/preset-typescript": "^7.8.3",
+        "@types/react": "^16.9.23",
+        "@types/react-dom": "^16.9.5",
+        "@types/react-router-dom": "^5.1.3",
         "babel-loader": "^8.0.6",
-        "babel-plugin-transform-object-rest-spread": "^6.26.0",
         "babel-preset-es2015": "^6.24.1",
         "babel-preset-es2017": "^6.24.1",
         "babel-preset-stage-0": "^6.24.1",
@@ -103,11 +107,10 @@
         "electron-packager": "^12.1.0",
         "file-system": "^2.2.2",
         "fork-ts-checker-webpack-plugin": "^4.1.0",
-        "google-protobuf": "^3.5.0",
+        "google-protobuf": "^3.11.4",
         "grpc": "^1.7.2",
         "grpc-loader": "^2.0.1",
         "prettier": "^1.19.1",
-        "react-hot-loader": "^3.1.3",
         "script-loader": "^0.7.2",
         "semantic-ui-react": "^0.77.0",
         "style-loader": "^0.19.0",
